+++ src/Toolkit.php
@@ -142,15 +142,14 @@ class Toolkit
 		$media_url = "";
 		$media_type = "";
 
-		$formats = $video_info->getFormats();
 		// find audio with highest quality
-		foreach($formats as $format)
+		foreach($video_info->getFormats() as $format)
 		{
-			if(strpos($format['type'], 'audio') !== false && intval($format['quality']) > intval($audio_quality))
+			if(strpos($format->getType(), 'audio') !== false && intval($format->getQuality()) > intval($audio_quality))
 			{
-				$audio_quality = $format['quality'];
-				$media_url = $format['url'];
-				$media_type = str_replace("audio/", "", $format['type']);
+				$audio_quality = $format->getQuality();
+				$media_url = $format->getUrl();
+				$media_type = str_replace("audio/", "", $format->getType());
 			}
 		}
 
@@ -171,8 +170,9 @@ class Toolkit
 				throw new Exception('MP3 downlod failed, no stream was found.');
 			}
 
-			$media_url = $formats[0]['url'];
-			$media_type = str_replace("audio/", "", $formats[0]['type']);
+			$fallbackFormat = $formats[0];
+			$media_url = $fallbackFormat->getUrl();
+			$media_type = str_replace("audio/", "", $fallbackFormat->getType());
 		}
 
 		$mp3dir = realpath($config->get('MP3TempDir'));
@@ -192,6 +192,7 @@ class Toolkit
 
 		// Download media from youtube success
 		$mp3Name = $_GET['title'] . '.mp3';
+
 		if($config->get('MP3Quality') !== "high" || $audio_quality === 0)
 		{
 			$audio_quality = intval($config->get('MP3Quality')) > intval($audio_quality) ? $audio_quality : $config->get('MP3Quality');